+++ example/augmentation/main.go
@@ -42,14 +42,18 @@ func tOne() {
 		panic(err)
 	}
 
-	device := gotch.CudaIfAvailable()
-	// device := gotch.CPU
+	// device := gotch.CudaIfAvailable()
+	device := gotch.CPU
 	imgTs := img.MustTo(device, true)
+	// h := imgTs.MustSize()[1]
+	// w := imgTs.MustSize()[2]
 
 	// t, err := aug.Compose(aug.WithRandomAutocontrast(1.0))
 	// t, err := aug.Compose(aug.WithRandomSolarize(aug.WithSolarizeThreshold(125), aug.WithSolarizePvalue(1.0)))
-	t, err := aug.Compose(aug.WithRandomAdjustSharpness(aug.WithSharpnessPvalue(1.0), aug.WithSharpnessFactor(10)))
+	// t, err := aug.Compose(aug.WithRandomAdjustSharpness(aug.WithSharpnessPvalue(1.0), aug.WithSharpnessFactor(10)))
 	// t, err := aug.Compose(aug.WithRandRotate(0, 360))
+	// Down sampling
+	// t, err := aug.Compose(aug.WithResize(h/2, w/2)) // NOTE. WithResize just works on CPU.
 	// t, err := aug.Compose(aug.WithResize(320, 320)) // NOTE. WithResize just works on CPU.
 	// t, err := aug.Compose(aug.WithRandomPosterize(aug.WithPosterizeBits(2), aug.WithPosterizePvalue(1.0)))
 	// t, err := aug.Compose(aug.WithRandomPerspective(aug.WithPerspectiveScale(0.6), aug.WithPerspectivePvalue(1.0)))
@@ -59,12 +63,13 @@ func tOne() {
 	// t, err := aug.Compose(aug.WithRandomVFlip(1.0))
 	// t, err := aug.Compose(aug.WithRandomHFlip(1.0))
 	// t, err := aug.Compose(aug.WithRandomEqualize(1.0))
-	// t, err := aug.Compose(aug.WithRandomCutout(aug.WithCutoutValue([]int64{124, 96, 255}), aug.WithCutoutScale([]float64{0.01, 0.1}), aug.WithCutoutRatio([]float64{0.5, 0.5})))
+	// t, err := aug.Compose(aug.WithRandomCutout(aug.WithCutoutValue([]int64{124, 96, 255}), aug.WithCutoutScale([]float64{0.01, 0.1}), aug.WithCutoutRatio([]float64{0.5, 0.5}), aug.WithCutoutPvalue(1.0)))
 	// t, err := aug.Compose(aug.WithCenterCrop([]int64{320, 320}))
 	// t, err := aug.Compose(aug.WithRandomAutocontrast())
 	// t, err := aug.Compose(aug.WithColorJitter(0.3, 0.3, 0.3, 0.3))
 	// t, err := aug.Compose(aug.WithGaussianBlur([]int64{5, 5}, []float64{1.0, 2.0}))
 	// t, err := aug.Compose(aug.WithRandomAffine(aug.WithAffineDegree([]int64{0, 15}), aug.WithAffineShear([]float64{0, 15})))
+	t, err := aug.Compose(aug.WithRandomAffine(aug.WithAffineDegree([]int64{0, 15}), aug.WithAffineTranslate([]float64{0.0, 0.1})))
 
 	out := t.Transform(imgTs)
 	fname := fmt.Sprintf("./bb-transformed.jpg")