+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed double free tensor at `vision/aug/function.go` Equalize func.
 - Changed `vision/aug` all input image should be `uint8` (Byte) dtype and transformed output has the same dtype (uint8) so that `Compose()` can compose any transformer options.
 - Fixed wrong result of `aug.RandomAdjustSharpness`
+- Fixed memory leak at `aug/function.getAffineGrid`
 
 ## [0.3.10]
 - Update installation at README.md